Fixed HTTP headers loaded out of order
When HTTP request was being built, the http headers were loaded out of
order.  The header that was used to build the RequestLine and URI
required the 'Host' header which was loaded in a subsequent iteration of
the While loop. This resulted in an exception being thrown when the URI
was created.  I resolved this by adding all of the other HTTP headers
first and then loading the RequestLineAndURI after the While loop
completed.
Also, the method that was used to add the headers to the
HeaderCollection in the RequestContext didn't have a constructor for the
HeaderCollection class and so a null exception was thrown as soon as a
header was attempted to add to the HeaderCollection.  I created a
default constructor for the RequestContext Class (Just as was done for
the ResponseContext Class) which solved this issue. I noticed that the
default code had a HeaderCollection constructor in the SetHeaders(),
however this method wasn't utilized.